Update es.hjson (Henry)
Hi @CalsiBotDev 
The main errors in the Spanish translation have been edited and corrected ,after almost 3 days of feedback. 
If there are any other details to be corrected, do not hesitate to write me here. In the case of Spanish language credits, I would like you to be able to link my user account with ID (𝓗𝓮𝓷𝓻𝔂) or username (@hfsoriad).
Regards,
Henry